[bug] Fix EnvVar to pick up valueFrom EnvVars
The code was setup to write any EnvVar provided in the Terraform Spec to
the Runner SetEnv function, so if you set valueFrom it would still look
for value and set that which would result in an empty string being set.

Tests added to verify that valueFrom is working with a Secret and a
ConfigMap valueFrom setting.

Signed-off-by: Luke Mallon (Nalum) <luke.mallon@weave.works>